fix(migration): align primary and foreign key types in RBAC schema

+++ database/migrations/1776870106281_create_user_roles_table.ts
@@ -6,8 +6,9 @@ export default class extends BaseSchema {
   async up() {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id')
+
       table
-        .bigInteger('user_id')
+        .integer('user_id')
         .unsigned()
         .notNullable()
         .references('id')
@@ -15,7 +16,7 @@ export default class extends BaseSchema {
         .onDelete('CASCADE')
 
       table
-        .bigInteger('role_id')
+        .integer('role_id')
         .unsigned()
         .notNullable()
         .references('id')
